pokedex/src/test/java/com/pokedex/pokedex/service/PokemonServiceTest.java
package com.pokedex.pokedex.service;

import static com.pokedex.pokedex.config.Constant.NAME_BULBASAUR;
import static com.pokedex.pokedex.config.Constant.NAME_IVYSAUR;
import static com.pokedex.pokedex.config.Constant.NUMBER_BULBASAUR;
import static com.pokedex.pokedex.config.Constant.NUMBER_IVYSAUR;
import static com.pokedex.pokedex.config.Constant.TYPE_BULBASAUR;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertNull;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.ArgumentMatchers.anyList;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.any;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import com.pokedex.pokedex.config.Constant;
import com.pokedex.pokedex.exception.PokemonNotFoundException;
import com.pokedex.pokedex.mapper.PokemonMapper;
import com.pokedex.pokedex.model.EvolutionDetail;
import com.pokedex.pokedex.model.Pokemon;
import com.pokedex.pokedex.model.PokemonPageResponse;
import com.pokedex.pokedex.model.PokemonResponse;
import com.pokedex.pokedex.repository.EvolutionDetailRepository;
import com.pokedex.pokedex.repository.JdbiTypeRepository;
import com.pokedex.pokedex.repository.PokemonRepository;
import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

import java.util.Optional;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockedStatic;
import org.mockito.Mockito;
import org.mockito.junit.jupiter.MockitoExtension;
import org.mockito.junit.jupiter.MockitoSettings;
import org.mockito.quality.Strictness;

@ExtendWith(MockitoExtension.class)
@MockitoSettings(strictness = Strictness.LENIENT) //ajuda a evitar falhas
public class PokemonServiceTest {
    @Mock
    private PokemonRepository pokemonRepository;

    @Mock
    private PokeApiService pokeApiService;

    @Mock
    private JdbiTypeRepository jdbiTypeRepository;

    @Mock
    private EvolutionDetailRepository evolutionDetailRepository;

    @InjectMocks
    private PokemonService pokemonService;

    @Test
    public void testAddNewPokemon_Success() throws Throwable {
        try (MockedStatic<PokemonMapper> pokemonMapperMockedStatic = Mockito.mockStatic(PokemonMapper.class)) {
            PokemonResponse pokemonResponse = new PokemonResponse();
            EvolutionDetail evolutionDetail1 = new EvolutionDetail();
            EvolutionDetail evolutionDetail2 = new EvolutionDetail();

            evolutionDetail1.setSelf(new Pokemon());
            evolutionDetail1.setEvolution(new Pokemon());
            evolutionDetail2.setSelf(new Pokemon());
            evolutionDetail2.setEvolution(new Pokemon());

            List<EvolutionDetail> evolutionDetailList = new ArrayList<>();
            evolutionDetailList.add(evolutionDetail1);
            evolutionDetailList.add(evolutionDetail2);

            when(pokeApiService.getPokemonNameOrNumber(NAME_BULBASAUR)).thenReturn(pokemonResponse);
            pokemonMapperMockedStatic.when(() -> PokemonMapper.toDomain(pokemonResponse)).thenReturn(evolutionDetailList);

            when(pokemonService.savePokemon(any(Pokemon.class))).thenAnswer(invocation -> invocation.getArguments()[0]);
            when(pokemonService.loadTypeFromDataBase(anyLong())).thenReturn(new ArrayList<>());
            when(evolutionDetailRepository.save(any(EvolutionDetail.class))).thenAnswer(invocation -> invocation.getArguments()[0]);

            List<EvolutionDetail> result = pokemonService.addNewPokemon(NAME_BULBASAUR);

            assertEquals(2, result.size());
            pokemonMapperMockedStatic.verify(() -> PokemonMapper.toDomain(pokemonResponse));
            verify(evolutionDetailRepository, times(2)).save(any(EvolutionDetail.class));

            for (EvolutionDetail evolutionDetail : result){
                assertNotNull(evolutionDetail.getSelf().getType());
                assertNotNull(evolutionDetail.getEvolution().getType());
            }
        }
    }

    @Test
    public void testListPokemons() {
        int page = 0;
        int pageSize = 10;
        int offSet = page * pageSize;

        EvolutionDetail evolutionDetail1 = new EvolutionDetail();
        Pokemon bulbasaur = new Pokemon();
        bulbasaur.setNumber(NUMBER_BULBASAUR);
        evolutionDetail1.setSelf(bulbasaur);

        EvolutionDetail evolutionDetail2 = new EvolutionDetail();
        Pokemon ivysaur = new Pokemon();
        ivysaur.setNumber(NUMBER_IVYSAUR);
        evolutionDetail2.setSelf(ivysaur);

        List<EvolutionDetail> evolutionDetails = new ArrayList<>();
        evolutionDetails.add(evolutionDetail1);
        evolutionDetails.add(evolutionDetail2);

        when(evolutionDetailRepository.findAll(pageSize, offSet)).thenReturn(evolutionDetails);
        when(evolutionDetailRepository.countAll()).thenReturn(20);

        //Mockando a PokemonMapper
        try (MockedStatic<PokemonMapper> pokemonMapperMockedStatic = Mockito.mockStatic(PokemonMapper.class)) {
            PokemonResponse pokemonResponse1 = new PokemonResponse();
            pokemonResponse1.setNumber(NUMBER_BULBASAUR);

            PokemonResponse pokemonResponse2 = new PokemonResponse();
            pokemonResponse2.setNumber(NUMBER_IVYSAUR);

            pokemonMapperMockedStatic.when(() -> PokemonMapper.toResponse(bulbasaur)).thenReturn(pokemonResponse1);
            pokemonMapperMockedStatic.when(() -> PokemonMapper.toResponse(ivysaur)).thenReturn(pokemonResponse2);

            List<EvolutionDetail> allEvolutions1 = Collections.emptyList();
            List<EvolutionDetail> allEvolutions2 = Collections.emptyList();

            when(evolutionDetailRepository.findBySelfNumber(NUMBER_BULBASAUR)).thenReturn(allEvolutions1);
            when(evolutionDetailRepository.findBySelfNumber(NUMBER_IVYSAUR)).thenReturn(allEvolutions2);

            pokemonMapperMockedStatic.when(() -> PokemonMapper.toResponseList(allEvolutions1)).thenReturn(Collections.emptyList());
            pokemonMapperMockedStatic.when(() -> PokemonMapper.toResponseList(allEvolutions2)).thenReturn(Collections.emptyList());

            when(jdbiTypeRepository.findByTypePokemonNumber(NUMBER_BULBASAUR)).thenReturn(TYPE_BULBASAUR);
            when(jdbiTypeRepository.findByTypePokemonNumber(NUMBER_IVYSAUR)).thenReturn(TYPE_BULBASAUR);

            PokemonPageResponse response = pokemonService.listPokemons(page, pageSize);

            assertEquals(2, response.getPokemons().size());
            assertEquals(page, response.getMeta().getPage());
            assertEquals(pageSize, response.getMeta().getPageSize());
            //TODO: necessário verificar se os valores abaixo estão invertidos
            //assertEquals(20, response.getMeta().getTotalElements());
            //assertEquals(2, response.getMeta().getTotalPage());

            for (PokemonResponse pokemonResponse : response.getPokemons()) {
                if (pokemonResponse.getNumber() == 1L) {
                    assertEquals(TYPE_BULBASAUR, pokemonResponse.getType());
                } else if (pokemonResponse.getNumber() == 2L) {
                    assertEquals(TYPE_BULBASAUR, pokemonResponse.getType());
                }
            }
        }
    }

    @Test
    public void testDeletePokemonByNameOrNumber() {
        Pokemon mockPokemon = new Pokemon();
        mockPokemon.setName(NAME_BULBASAUR);
        mockPokemon.setNumber(NUMBER_BULBASAUR);

        when(pokemonRepository.getPokemonByNameOrNumber(mockPokemon.getName())).thenReturn(Optional.of(mockPokemon));

        pokemonService.deletePokemonByNameOrNumber(mockPokemon.getName());

        verify(evolutionDetailRepository, times(1)).deleteBySelfNumber(mockPokemon.getNumber());
        verify(pokemonRepository, times(1)).deletePokemon(mockPokemon.getNumber());
    }

    @Test
    public void testDeletePokemonByNameOrNumber_NotFound() {
        String nameOrNumber = "XXXXXX";

        when(pokemonRepository.getPokemonByNameOrNumber(nameOrNumber)).thenReturn(Optional.empty());

        assertThrows(PokemonNotFoundException.class, () -> {
            pokemonService.deletePokemonByNameOrNumber(nameOrNumber);
        });
    }

    @Test
    public void testGetEvolutionsByPokemonNumber() {
        EvolutionDetail evolutionDetail1 = new EvolutionDetail();
        Pokemon pokemon1 = new Pokemon();
        pokemon1.setNumber(NUMBER_BULBASAUR);
        pokemon1.setName(NAME_BULBASAUR);
        evolutionDetail1.setSelf(pokemon1);

        EvolutionDetail evolutionDetail2 = new EvolutionDetail();
        Pokemon pokemon2 = new Pokemon();
        pokemon2.setNumber(NUMBER_IVYSAUR);
        pokemon2.setName(NAME_IVYSAUR);
        evolutionDetail2.setSelf(pokemon2);

        List<EvolutionDetail> evolutionDetails = new ArrayList<>();
        evolutionDetails.add(evolutionDetail1);
        evolutionDetails.add(evolutionDetail2);

        when(evolutionDetailRepository.findBySelfNumber(NUMBER_BULBASAUR)).thenReturn(evolutionDetails);

        try (MockedStatic<PokemonMapper> pokemonMapperMockedStatic = Mockito.mockStatic(PokemonMapper.class)){
            PokemonResponse pokemonResponse1 = new PokemonResponse();
            pokemonResponse1.setNumber(NUMBER_BULBASAUR);
            pokemonResponse1.setName(NAME_BULBASAUR);

            PokemonResponse pokemonResponse2 = new PokemonResponse();
            pokemonResponse2.setNumber(NUMBER_IVYSAUR);
            pokemonResponse2.setName(NAME_IVYSAUR);

            pokemonMapperMockedStatic.when(() -> PokemonMapper.toResponse(pokemon1)).thenReturn(pokemonResponse1);
            pokemonMapperMockedStatic.when(() -> PokemonMapper.toResponse(pokemon2)).thenReturn(pokemonResponse2);

            List<PokemonResponse> pokemonResponses = new ArrayList<>();
            pokemonResponses.add(pokemonResponse2);
            pokemonMapperMockedStatic.when(() -> PokemonMapper.toResponseList(anyList())).thenReturn(pokemonResponses);

            //TODO: as evolutions estão vindo null, necessário corrigir
            PokemonResponse response = pokemonService.getEvolutionsByPokemonNumber(NUMBER_BULBASAUR);

            assertNotNull(response);
            assertEquals(NUMBER_BULBASAUR, response.getNumber());
            assertEquals(NAME_BULBASAUR, response.getName());
            assertEquals(2, response.getEvolutions().size());
            assertEquals(NUMBER_IVYSAUR, response.getEvolutions().get(0).getNumber());
            assertEquals(NAME_IVYSAUR, response.getEvolutions().get(0).getName());

            verify(evolutionDetailRepository).findBySelfNumber(NUMBER_BULBASAUR);
        }
    }

    @Test
    public void testGetEvolutionsByPokemonNumber_NoEvolutions() {
        when(evolutionDetailRepository.findBySelfNumber(Constant.NUMBER_BULBASAUR)).thenReturn(Collections.emptyList());

        PokemonResponse response = pokemonService.getEvolutionsByPokemonNumber(Constant.NUMBER_BULBASAUR);

        assertNull(response);
    }
}